requirements.txt
streamlit
yt-dlp
openai-whisper
torch
ffmpeg-python
kokoro
soundfile
openai
pymupdf
python-pptx
python-docx
python-dotenv
